Move CLI to own file
