+++ src/bin/why/main.rs
@@ -1,68 +1,20 @@
 extern crate pest;
 extern crate y_lang;
 
+mod cli;
+
+use cli::*;
+
 use std::{collections::HashMap, error::Error, fs};
 
-use clap::{Parser as CParser, ValueEnum};
 use log::{error, info};
 use y_lang::{
     compiler::Compiler,
     loader::{load_module, load_modules, Module, Modules},
 };
 
-#[derive(CParser, Debug)]
-#[command(author, version, about)]
-struct Cli {
-    /// The path to the why source file.
-    #[arg(index = 1)]
-    file: std::path::PathBuf,
-
-    /// Whether to dump the parsed AST (for debugging).
-    #[arg(long)]
-    dump_parsed: bool,
-
-    /// Whether to dump the type-checked AST (for debugging).
-    #[arg(long)]
-    dump_typed: bool,
-
-    /// The path to the output binary.
-    #[arg(short, long)]
-    output: Option<std::path::PathBuf>,
-
-    /// Specify the log level of the compiler.
-    #[arg(value_enum, short, long, default_value_t = LogLevel::default())]
-    verbosity: LogLevel,
-}
-
-#[derive(ValueEnum, Clone, Default, Debug)]
-enum LogLevel {
-    #[default]
-    #[value(alias("0"))]
-    Error,
-
-    #[value(alias("1"))]
-    Warn,
-
-    #[value(alias("2"))]
-    Info,
-
-    #[value(alias("3"))]
-    Debug,
-}
-
-impl From<LogLevel> for log::Level {
-    fn from(value: LogLevel) -> Self {
-        match value {
-            LogLevel::Error => log::Level::Error,
-            LogLevel::Warn => log::Level::Warn,
-            LogLevel::Info => log::Level::Info,
-            LogLevel::Debug => log::Level::Debug,
-        }
-    }
-}
-
 fn main() -> Result<(), Box<dyn Error>> {
-    let args = Cli::parse();
+    let args = Cli::init();
 
     simple_logger::init_with_level(args.verbosity.into()).unwrap();
